Early version of test program
The test program can be compiled in two mode (for now using two different
Makefiles).
 - Makefile: to control the actual hardware
 - Makefile.virtual: to control virtual hardware
The second version can be run on a regular (desktop) machine. It contains
some rough checks about the way how the YM2413 pins are used. It also prints
some higher level info based on how the YM2413 pins are being toggled (e.g.
it prints when which registers are written). So in a way this can be used
to test the test programm.