Added tool2 subdirectory
For the moment this is still very close to the 'tool' subdir. I made a copy
instead of modifying that dir so that I always have a working fallback in case
I screw something up.

The only real change so far is that I split the various '.hh' files in 'actual'
header and  implementation files (.hh + .cc). (Before it were actually also
implementation files that were #included in main.cc).