Proper synchronization of timer-1
Timer-1 determines which of the 72 samples in the YM2413 period we sample.
Before we used a hack to sync this. And that hack only worked after resetting
the board, not after reprogramming and then running the program (without
reset).

Now we have a routine that properly detects the position of music channel 1 and
synchronizes timer-1 to that position.