More dynamic firmware / calibration tool
Made the firmware more dynamic, meaning that we don't need to reflash a new
program each time we want to test a different thing. Instead we can now send
instructions over USB (like write YM2413 registers or start/stop capturing).

Also added a tool 'calibrate' that uses this new firmware to capture a sine
wave and based on this calculates the positions of the YM2413 output levels.